# openapi.yaml
openapi: 3.0.0
info:
  title: MiniBlog API
  description: API REST para gestionar autores y publicaciones de MiniBlog.
  version: 1.0.0
  contact:
    name: Felipe Salazar
    email: felipesalazarcruz4@gmail.com
servers:
  - url: http://localhost:3000
    description: Servidor local
paths:
...
  /comments:
    get:
      summary: Listar todos los comentarios
      responses:
        "200":
          description: Lista de comentarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      summary: Crear un nuevo comentario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comentario creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          description: Datos inválidos o post_id/author_id inexistentes
  /comments/{id}:
    put:
      summary: Actualizar el contenido de un comentario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
      responses:
        "200":
          description: Comentario actualizado
        "400":
          description: Contenido inválido
        "404":
          description: Comentario no encontrado
    delete:
      summary: Eliminar un comentario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Comentario eliminado
        "404":
          description: Comentario no encontrado
  /comments/post/{postId}:
    get:
      summary: Listar comentarios de un post específico, con nombre del autor incluido
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lista de comentarios del post
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommentWithAuthor"
components:
  schemas:
...
    Comment:
      type: object
      properties:
        id:
          type: integer
        post_id:
          type: integer
        author_id:
          type: integer
        content:
          type: string
        created_at:
          type: string
          format: date-time
    CommentInput:
      type: object
      required:
        - post_id
        - author_id
        - content
      properties:
        post_id:
          type: integer
        author_id:
          type: integer
        content:
          type: string
    CommentWithAuthor:
      allOf:
        - $ref: "#/components/schemas/Comment"
        - type: object
          properties:
            author_name:
              type: string